Refactor #35 - 테스트 폴더 구조 변경
