+++ src/test/java/com/maruhxn/boardserver/config/RestDocsConfiguration.java
@@ -1,4 +1,4 @@
-package com.maruhxn.boardserver;
+package com.maruhxn.boardserver.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
