+++ src/test/java/com/maruhxn/boardserver/controller/AuthControllerTest.java
@@ -1,6 +1,6 @@
 package com.maruhxn.boardserver.controller;
 
-import com.maruhxn.boardserver.TestSupport;
+import com.maruhxn.boardserver.support.TestSupport;
 import com.maruhxn.boardserver.common.exception.ErrorCode;
 import com.maruhxn.boardserver.dto.request.auth.LoginRequest;
 import com.maruhxn.boardserver.dto.request.auth.RegisterRequest;
