+++ src/test/java/com/maruhxn/boardserver/controller/PostControllerTest.java
@@ -1,6 +1,6 @@
 package com.maruhxn.boardserver.controller;
 
-import com.maruhxn.boardserver.TestSupport;
+import com.maruhxn.boardserver.support.TestSupport;
 import com.maruhxn.boardserver.common.exception.ErrorCode;
 import com.maruhxn.boardserver.domain.Post;
 import com.maruhxn.boardserver.dto.request.posts.CreatePostRequest;
